Create README.md
cat_a_chrome_extension
======================

Catify the Internet

Ever read the news and think to yourself, 'hey, this is depressing.'  Ever think to yourself, ' I love this article, but I wish
it was in Comic Sans.  We have a solution. 

This Chrome Extension will take that frown and turn it upside down. 

/////////////////////////////

implementation 

DOWNLOAD THE SOURCE FILES

Visit chrome://extensions in your browser (or open up the Chrome menu by clicking the icon to the far right of the Omnibox: The menu's icon is three horizontal bars.. and select Extensions under the Tools menu to get to the same place).

Ensure that the Developer mode checkbox in the top right-hand corner is checked.

Click Load unpacked extension… to pop up a file-selection dialog.

Navigate to the directory in which your extension files live, and select it.